fix scorelabel

+++ ScoreLabel.cs
@@ -3,9 +3,9 @@ using System;
 
 public partial class ScoreLabel : Label
 {
-    public float score = 0;
+    public double score = 0;
     bool PlayerLost = false;
-    public override void _Process(float delta)
+    public override void _Process(double delta)
     {
         if (PlayerLost) { return; }
         score += delta;
